Change default network for examples to (null) which defaults to the administrative zone of 239.192.0.1. Fix warnings / errors on 64-bit shifting on 32-bit builds.

+++ trunk/openpgm/pgm/include/pgm/sn.h
@@ -29,7 +29,7 @@ enum {
     PGM_UINT32_SIGN_BIT = (1<<31)
 };
 
-#define PGM_UINT64_SIGN_BIT ((guint64)1<<63)
+#define PGM_UINT64_SIGN_BIT (1ULL<<63)
 
 /* 32 bit */
 static inline gboolean pgm_uint32_lt (guint32 s, guint32 t)
